Added _query_repr methods and changed builder to use them
Added _query_repr methods and changed builder to use them.
Also added a property for the DB name. Since Conn Strings can vary a lot
(for example, Oracle TNS names), we just use the DB name used in the
ctor as the "name". It can be a config name or the full conn string.