Simplified the formatters
Removed support (which at some point stopped working, BTW) for things
other than "list of namedtuples". It should be easier now to merge these
three methods in an object (ResultSet, see issue 17). Even if we don't
go that route, it's still not worth it to make everything harder to just
print a list. No one will use the formatters outside of pyquebec.